+++ INFO.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright 2020 Xpp521
+# Copyright (C) 2020 Xpp521
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -14,19 +14,9 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-"""
-This module will automatically import "Strings" according to the setting.
-If the setting doesn't exist, use the system default language.
-"""
-from PyQt5.QtCore import QSettings
-from locale import getdefaultlocale
-from importlib import import_module
-lang = QSettings().value('Language/data')
-if lang:
-    module = import_module('.Strings_{}'.format(lang), 'Strings')
-    from . import module as Strings
-else:
-    if 'zh_cn' in getdefaultlocale()[0].lower():
-        from . import Strings_cn as Strings
-    else:
-        from . import Strings_en as Strings
+VERSION = 'v1.2.0'
+AUTHOR_NAME = 'Xpp'
+APP_NAME = 'XSearch'
+AUTHOR_EMAIL = 'Xpp233@foxmail.com'
+OPEN_SOURCE_ADDRESS = 'https://github.com/Xpp521/XSearch'
+RELEASE_ADDRESS = 'https://github.com/Xpp521/XSearch/releases'