+++ TrayIcon.py
@@ -38,7 +38,7 @@ class TrayIcon(QSystemTrayIcon):
     def reload_ui(self, text=True, qss=True):
         if text:
             self.setIcon(QIcon(join('Icons', 'XSearch.ico')))
-            from Strings import Strings
+            from Languages import Strings
             self.setToolTip(Strings.APP_NAME)
             self.actions.get('search').setText(Strings.TRAY_SEARCH)
             self.actions.get('setting').setText(Strings.TRAY_SETTING)